Clean-up based on feedback
