+++ src/main/java/org/omnifaces/cdi/pooled/PooledContext.java
@@ -2,7 +2,10 @@ package org.omnifaces.cdi.pooled;
 
 import java.lang.annotation.Annotation;
 import java.util.Map;
+import java.util.Queue;
 import java.util.concurrent.ConcurrentHashMap;
+import java.util.concurrent.ConcurrentLinkedQueue;
+import java.util.stream.IntStream;
 
 import javax.enterprise.context.spi.Context;
 import javax.enterprise.context.spi.Contextual;
@@ -27,6 +30,7 @@ public class PooledContext implements Context {
 		if (contextual instanceof Bean) {
 
 			try {
+				@SuppressWarnings("unchecked")
 				Class<T> proxyClass = ((Class<T>) proxyClasses.computeIfAbsent(contextual, ctx -> {
 					ProxyFactory factory = new ProxyFactory();
 
@@ -63,15 +67,63 @@ public class PooledContext implements Context {
 		return true;
 	}
 
+	@SuppressWarnings("unchecked")
 	public <T> PoolKey<T> allocateBean(Contextual<T> contextual) {
 		return ((InstancePool<T>) instancePools.computeIfAbsent(contextual, InstancePool::new)).allocateInstance();
 	}
 
+	@SuppressWarnings("unchecked")
 	public <T> void releaseBean(PoolKey<T> poolKey) {
 		((InstancePool<T>) instancePools.get(poolKey.getContextual())).releaseInstance(poolKey);
 	}
 
+	@SuppressWarnings("unchecked")
 	public <T> T getBean(PoolKey<T> poolKey, CreationalContext<T> creationalContext) {
 		return ((InstancePool<T>) instancePools.get(poolKey.getContextual())).getInstance(poolKey, creationalContext);
+	}
+
+	static class InstancePool<T> {
+
+		private static final int MAX_NUMBER_OF_INSTANCES = 10;
+
+		private final Contextual<T> contextual;
+		private final Map<PoolKey<T>, T> instances = new ConcurrentHashMap<>();
+		private final Queue<PoolKey<T>> freeInstanceKeys = new ConcurrentLinkedQueue<>();
+
+		public InstancePool(Contextual<T> contextual) {
+			this.contextual = contextual;
+
+			IntStream.range(0, MAX_NUMBER_OF_INSTANCES)
+			         .mapToObj(i -> new PoolKey<>(contextual, i))
+			         .forEach(freeInstanceKeys::add);
+		}
+
+		public T getInstance(PoolKey<T> poolKey) {
+			if (!poolKey.getContextual().equals(contextual)) {
+				throw new IllegalArgumentException();
+			}
+
+			return instances.get(poolKey);
+		}
+
+		public T getInstance(PoolKey<T> poolKey, CreationalContext<T> context) {
+			if (!poolKey.getContextual().equals(contextual)) {
+				throw new IllegalArgumentException();
+			}
+
+			return instances.computeIfAbsent(poolKey, key -> contextual.create(context));
+		}
+
+		public PoolKey<T> allocateInstance() {
+			return freeInstanceKeys.remove();
+		}
+
+		public void releaseInstance(PoolKey<T> key) {
+			if (!contextual.equals(key.getContextual())) {
+				throw new IllegalArgumentException();
+			}
+
+			freeInstanceKeys.add(key);
+		}
 	}
 }
